Align with the new Base32H spec
Treating uppercase and lowercase L separately is probably a bit too
confusing.  L/l is still its own digit independent from 1; this might
result in some ambiguity in lowercase identifiers, but given that
the canonical digits are uppercase I don't expect this to be an
issue in practice (people reading existing identifiers would hopefully
be reading the canonical/uppercase representations, given that those
are the ones compliant encoders emit, so accepting lowercase
identifiers is more for data entry convenience).